Repackage bindings in sample app
