--- samples/android-kotlin/src/main/java/dev/arunkumar/scabbard/debug/DebugBindings.kt
@@ -1,90 +1,0 @@
-package dev.arunkumar.scabbard.debug
-
-import android.app.Application
-import dagger.Binds
-import dagger.Module
-import dagger.Provides
-import dagger.multibindings.IntoSet
-import javax.inject.Inject
-import javax.inject.Named
-import javax.inject.Qualifier
-import javax.inject.Singleton
-
-class MemberInjectionTarget
-
-class UnScopedBinding @Inject constructor()
-
-@Singleton
-class SimpleSingleton @Inject constructor()
-
-@Singleton
-class ComplexSingleton
-@Inject
-constructor(
-  private val unScopedBinding: UnScopedBinding,
-  private val simpleSingleton: SimpleSingleton,
-  private val application: Application
-)
-
-class ProvisionBinding
-
-@Module
-object ProvisionModule {
-  @Provides
-  @Singleton
-  fun providesProvisionBinding() = ProvisionBinding()
-}
-
-interface MultiBindingType
-
-class SimpleMultiBindingType @Inject constructor() : MultiBindingType
-
-class ComplexMultiBindingType
-@Inject
-constructor(val provisionBinding: ProvisionBinding) : MultiBindingType
-
-@Module
-abstract class MultiBindingsProvisionModule {
-  @Binds
-  @IntoSet
-  abstract fun simpleMultiBindingType(simpleMultiBindingType: SimpleMultiBindingType): MultiBindingType
-
-  @Binds
-  @IntoSet
-  abstract fun complexMultiBindingType(complexMultiBindingType: ComplexMultiBindingType): MultiBindingType
-}
-
-interface DelegateBinding
-
-class DefaultDelegateBinding
-@Inject
-constructor(val unScopedBinding: UnScopedBinding) : DelegateBinding
-
-@Module
-abstract class DelegateBindingModule {
-  @Binds
-  abstract fun bindsDelegateBinding(defaultDelegateBinding: DefaultDelegateBinding): DelegateBinding
-}
-
-class NamedBinding
-
-@Module
-object NamedProvisionModule {
-  @Provides
-  @Named("named")
-  @Singleton
-  fun providesNamedBinding() = NamedBinding()
-}
-
-@Qualifier
-@Retention
-annotation class SimpleQualifier
-
-class QualifiedBinding
-
-@Module
-object QualifiedProvisionModule {
-  @SimpleQualifier
-  @Provides
-  fun providesQualifiedBinding() = QualifiedBinding()
-}
