+++ samples/android-kotlin/src/main/java/dev/arunkumar/scabbard/di/AppComponent.kt
@@ -6,7 +6,6 @@ import dagger.Component
 import dagger.android.AndroidInjector
 import dagger.android.support.AndroidSupportInjectionModule
 import dev.arunkumar.scabbard.App
-import dev.arunkumar.scabbard.debug.*
 import dev.arunkumar.scabbard.home.HomeActivity
 import dev.arunkumar.scabbard.home.simple.SimpleSubcomponent
 import javax.inject.Singleton
@@ -27,7 +26,7 @@ import javax.inject.Singleton
 )
 interface AppComponent : AndroidInjector<App> {
 
-  fun inject(memberInjectionTarget: MemberInjectionTarget)
+  fun inject(memberInjectionBinding: MemberInjectionBinding)
 
   fun simpleSubcomponentFactory(): SimpleSubcomponent.Factory
