--- samples/android-kotlin/src/main/java/dev/arunkumar/scabbard/di/Util.kt
@@ -1,6 +1,0 @@
-package dev.arunkumar.scabbard.di
-
-import android.app.Application
-import dev.arunkumar.scabbard.App
-
-val Application.appComponent get() = (this as App).appComponent
